deps(package.json): Bump `@typescript-eslint/*` to "^5.58.0"
This is to get rid of the below warnings:

```
70% sealing finish module graph ESLintWebpackPlugin_1=============

WARNING: You are currently running a version of TypeScript which is not officially supported by @typescript-eslint/typescript-estree.

You may find that it works just fine, or you may not.

SUPPORTED TYPESCRIPT VERSIONS: >=3.3.1 <5.0.0

YOUR TYPESCRIPT VERSION: 5.0.4

Please only submit bug reports when using the officially supported version.
```

Ref: https://github.com/typescript-eslint/typescript-eslint/pull/6662